Refactor sealed secrets.
