+++ lib/edge-deploy.js
@@ -4,21 +4,16 @@
  * Copyright 2023 AMRC
  */
 
-import child_process from "child_process";
-import fs from "fs";
-import fs_p from "fs/promises";
-import path from "path";
 import stream from "stream";
 
-import concat_stream from "concat-stream";
 import express from "express";
-import tmpfile from "tmp-promise";
 
 import { Debug, UUIDs } from "@amrc-factoryplus/utilities";
 
-import { Checkout } from "./checkout.js";
-import { Git, Edge } from "./uuids.js";
-import * as manifests from "./manifests.js";
+import { Checkout }         from "./checkout.js";
+import { SealedSecrets }    from "./secrets.js";
+import * as manifests       from "./manifests.js";
+import { Git, Edge }        from "./uuids.js";
 
 const FLUX_NS = "flux-system";
 
@@ -28,12 +23,12 @@ export class EdgeDeploy {
     constructor (opts) {
         this.fplus      = opts.fplus;
         this.http_url   = opts.http_url;
-        this.git_email  = opts.git_email;
         this.group      = opts.repo_group;
-        this.cert_dir   = opts.cert_dir;
 
         this.log = debug.log.bind(debug, "edge");
         this.routes = express.Router();
+
+        this.secrets = new SealedSecrets(opts);
     }
 
     wrap (fn) {
@@ -186,104 +181,32 @@ export class EdgeDeploy {
         await co.commit("Written flux source manifests.");
     }
 
-    async write_sealed_secret (co, x509, s_req) {
-        const { namespace, name, key, content } = s_req;
-
-        const sealed = await this.run_kubeseal(x509, s_req);
-        this.log("Got sealed data: %s", sealed);
-
-        const obj = await co.read_manifest(namespace, "SealedSecret", name) ??
-            manifests.sealed_secret(namespace, name);
-        obj.spec.encryptedData[key] = sealed;
-        await co.write_manifest(obj);
-        await co.commit("Updated sealed secret %s/%s/%s.", 
-            namespace, name, key);
-    }
-
-    async _seal_secret (cluster, s_req) {
-        const info = await this.fplus.ConfigDB
-            .get_config(Edge.App.Cluster, cluster);
-        if (!info) return 404;
-
-        const x509 = info.kubeseal_cert;
-        if (!x509) return 503;
-
-        const co = await Checkout.clone({ fplus: this.fplus, url: info.flux });
-        await this.write_sealed_secret(co, x509, s_req);
-        if (s_req.dryrun)
-            await co.dispose()
-        else
-            await co.push();
-
-        return 204;
-    }
-
     async seal_secret (req, res) {
-        const { cluster, namespace, name, key } = req.params;
-        const s_req = { 
-            namespace, name, key, 
+        const opts = { 
+            ...req.params,
             content:    req, 
             dryrun:     "dryrun" in req.query,
         };
         
         const ok = await this.fplus.Auth.check_acl(
-            req.auth, Edge.Perm.Secrets, cluster, true);
+            req.auth, Edge.Perm.Secrets, opts.cluster, true);
         if (!ok) return res.status(403).end();
 
-        const st = await this._seal_secret(cluster, s_req);
+        const st = await this.secrets.seal_secret(opts);
         res.status(st).end();
     }
 
     async delete_sealed_secret (req, res) {
-        const { cluster, namespace, name, key } = req.params;
-        const dryrun = "dryrun" in req.query;
-
-        const ok = await this.fplus.Auth.check_acl(
-            req.auth, Edge.Perm.Secrets, cluster, true);
-        if (!ok) return res.status(403).end();
-
-        const info = await this.fplus.ConfigDB
-            .get_config(Edge.App.Cluster, cluster);
-        if (!info) return res.status(404).end();
-
-        const mf = [namespace, "SealedSecret", name];
-        const co = await Checkout.clone({ fplus: this.fplus, url: info.flux });
-        const obj = await co.read_manifest(...mf);
-        if (obj) {
-            const enc = obj.spec.encryptedData;
-            delete enc[key];
-            if (Object.keys(enc).length == 0)
-                await co.unlink_manifest(...mf);
-            else
-                await co.write_manifest(obj);
-            if (dryrun)
-                await co.dispose();
-            else
-                await co.push("Remove sealed secret %s/%s/%s.",
-                    namespace, name, key);
+        const opts = {
+            ...req.params,
+            dryrun:         "dryrun" in req.query,
         }
 
-        return res.status(204).end();
+        const ok = await this.fplus.Auth.check_acl(
+            req.auth, Edge.Perm.Secrets, opts.cluster, true);
+        if (!ok) return res.status(403).end();
+
+        const st = await this.secrets.delete_secret(opts);
+        res.status(st).end();
     }
-
-    run_kubeseal (x509, s_req) {
-        return tmpfile.withFile(async ({ path: cert }) => {
-            await fs_p.writeFile(cert, x509);
-
-            const sealed = await new Promise(resolve => {
-                const kid = child_process.spawn("/usr/local/bin/kubeseal", [
-                    "--cert", cert,
-                    "--namespace", s_req.namespace,
-                    "--name", s_req.name,
-                    "--raw"]);
-                s_req.content.pipe(kid.stdin);
-                kid.stderr.pipe(concat_stream(
-                    { encoding: "string" }, 
-                    err => err && this.log("Kubeseal error: %s", err)));
-                kid.stdout.pipe(concat_stream({ encoding: "string" }, resolve));
-            });
-
-            return sealed;
-        }, { tmpdir: this.cert_dir, prefix: "cert-" });
-    };
 }
